Move parsing of lockfile into it's own class, make an interface for it
