+++ app/Parse/Parse.php
@@ -1,0 +1,7 @@
+<?php
+namespace App\Parse;
+
+interface Parse
+{
+    public function get_coordinates($packages) : array;
+}
